requirements.txt
pyserial==3.5
flask==1.1.2
pycrypto==2.6.1
Flask-SQLAlchemy==2.5.1
flask-login=0.5.0